bugfix #1091659
In the function addCharToTable, I guess there's a
problem: you do 'p = findCodeValue(c, font->codeTable,
0, font->nGlyphs);' which returns the index p of
font->codeTable where the character c is. And if c is
not found, then it returns the index where to add it
(that is, at the end of font->codeTable).

Then, you make a test to really knows if the character
is at the 'p' position:
'if (font->codeTable != NULL && font->codeTable[p] == c)'
If it is found, the function returns. But this may be
bad because as codeTable has 'p' elements (that is, the
last index is 'p-1'), you try to read somewhere in the
memory.
And thus, if the test comes true, the function returns
without inserting the new character.

I guess you can correct this by adding another test:
'if (font->codeTable != NULL && p != font->nGlyphs &&
font->codeTable[p] == c)'. Like that, the character is
added.

Submitted by: T JuJu