bugfix: http://bugs.libming.net/show_bug.cgi?id=52
this bug is not related to the as-grammar. your example triggered an other
issue. calling swf5error() in the action part of a bison file does not set the
parser in an error state. Therefore swf5parse() returns success. This issues
occurs only when fiddling with mings error reporting methods (like makeswf).