Un-define DEBUGSTACK (util/decompile.c)
pop(), peek() and co. are designed to crash whenever the stack becomes
NULL but this behavior is currently short-circuited by the DEBUGSTACK
option, a debug option which allows further processing when Stack == NULL.

With DEBUGSTACK defined, whenever the stack becomes NULL it is replaced
by a dummy element like "// *** pop(): INTERNAL STACK ERROR FOUND ***".

While this is an acceptable feature for debugging purposes, this is not
something we want for production because it may lead other functions to
crash (infinite loops, buffer over reads...) in an especially undebuggabble
way.