Fix code duplication issue in listfdb
listfdb is shipping a large portion of duplicate code from the read
module. This is a major security flaw given that the copied source
code is affected by a large number of security issues (actually,
almost all issues discovered in read.c for a long long time).

In this patch we remove the duplicate code from listfdb.

In order to do this we:
- Link again read.h
- Add the required dependency in util/Makefile.am
- Make the bufbits variable extern in read.h

This patch fixes #107 and #106 (and probably a *lot* of other issues).