Fix heap-buffer-overflows when accessing pool
Constants are usually retrieved from the constant pool without verifying
that the pool actually contains them, which may lead to various heap
buffer overflow issues.

In this patch we add a counter keeping track of how many elements the pool
contains, and checks making sure that whenever the pool is accessed, the
constant in present in the pool (constant position < pool counter).

Also, do not return "" when a pointer is excepted (it should be legal to free
this return value).

This patch fixes #112 (CVE-2018-7875), fixes #120 (CVE-2018-7871),
fixes #117 (CVE-2018-7870), fixes #114 (CVE-2018-7872), fixes #122,
fixes #113 (CVE-2018-7868), fixes #123.