Fix null pointer dereference in getName/getString
Whenever getString or getName are called with an act such that act->p.String
is a NULL pointer, a NULL pointer dereference might happen
(strlen(act->p.string) is called).

In this commit we add checks at the beginning of the PUSH_STRING block so
that a warning is displayed and an empty string is returned in this case.

This patch fixes #121.