Perform deep copy in pushdup (instead of shallow)
Until now, the element duplication in pushdup was performed via
t->val = Stack->val.

While this is perfectly fine for integer/double/register values,
this may create nasty, hard to debug issues with Strings. In fact,
when called with a String at the top of the stack, pushdup would
only push *a reference* to the same String element (shallow copy),
later allowing to modify several stack elements at once, which may
potentially lead to NULL pointer dereferences or any other
unspecified impact.

In this patch we implement deep copy in pushdup:
* If the type of the stack element is 's' (for String), we
  allocate a new buffer and copy the String into it.
* Otherwise we simply proceed as before, that is we do t->val = Stack->val
  which is perfectly fine since we are not dealing with pointers.

This patch is the last part of the patch for #121 (fixes #121), which
should now be completely fixed.