decompile: Fix null pointer dereference in getInt
When getInt is passed a PUSH_REGISTER parameter, it retrieves the
content of this register and returns the value contained by this
register as an int. When this register is empty, we call getInt with
a NULL pointer and a null pointer dereference occurs.

In this patch we first make sure that regs[act->p.RegisterNumber] is
not NULL before doing anything with it.

Fixes #133 (CVE-2018-9132).