swftypes: fix type issue causing memory exhaustion
This commit fixes the memory exhaustion issue in
parseSWF_ACTIONRECORD (fixes: #109, CVE-2018-7876).

The original issue consists is triggered by an integer overflow in
parseSWF_ACTIONRECORD, where we read a UI16 and store it in a WORD,
which is defined as SI16. This is because type WORD (=SI16) is used
for NumParam (in SWF_ACTIONDEFINEFUNCTION), while the specification
says it should be UI16 (page 92 of the spec).

This patch addresses this type issue by changing type of NumParam
from WORD to UI16.