decompile: fix null pointer dereference in newVar3
getString (indirectly called by getName) is passed a variable of non
standard type 10 (= "PUSH_VARIABLE"), which seems to return the
string contained in passed variable, without quotes. If contained
string is NULL, a NULL pointer is returned, which later causes NULL
pointer dereference.

In this patch we address this issue such that if the variable contains
an invalid string, we act just like in the PUSH_STRING case. Otherwise
a copy of the string is returned.

Fixes: #118 (CVE-2018-7866).