decompile: introduce new method Offset
The getString method in decompile.c is vulnerable to a buffer
overflow which can be triggered using a crafted SWF file.

This vulnerability is the consequence of unchecked accesses to the
actions array when getting the offset of SWF_ACTIONRECORD objects.

This pattern is present a bit everywhere in the source code, leading
to a large number of potential flaws similar to this one. In this
commit we introduce a new Offset method similar to the OpCode method
which handles bound checking when retrieving the offset of
SWF_ACTIONRECORD objects.

This commit also modifies getString to use this newly introduced
method and address the previously explained bug.

Usage of the newly introduced Offset method will be generalized in a
future commit.

Please, note that this commit won't be sufficient to fix #144
(CVE-2018-11226) since another issue is triggered by the same sample.