decompile: fix loop cond issue leading to OOB read
In decompileSETTARGET a while loop is used to count the number of
operations until a certain type of operation has been reached. This
loop uses action_cnt+n < maxn as stop condition, meaning that
action_cnt+n = maxn might be true after the loop.

This is wrong because action_cnt is used as the number of operations
to process in an array of maxn-n-1 elements.

Fix the loop's stop condition and switch to for loop for better
readability.

This patch is the second part of the CVE-2018-11226 fix (fixes: #144).